add timers classe
